Refactor theme selector Fixes #121
