+++ apps/web/src/components/app-sidebar/app-sidebar.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { ModeToggle } from "@/components/mode-toggle/mode-toggle";
 import {
   Sidebar,
   SidebarContent,
@@ -12,7 +13,6 @@ import {
 import { ScenarioData } from "@/models/scenario";
 import { useParams } from "next/navigation";
 import { ScenarioItem } from "./scenarioItem";
-import { ModeToggle } from "./mode-toggle";
 
 // Extend the props from the base Sidebar and add scenarios
 interface AppSidebarProps extends React.ComponentProps<typeof Sidebar> {
